Fixes for 64-bit systems
1) config/intlen.c tested int, but should have tested long.

2) Src/MMTK_trajectory.c needs to use int instead of long when
   interpreting the variable "step" read from the trajectory file.

+++ config/intlen.c
@@ -3,5 +3,5 @@
 
 int main(int argc, char **argv)
 {
-  printf("%d\n", sizeof(int));
+  printf("%d\n", sizeof(long));
 }
